Prune some more unused methods, work in a few more constants.
The Asc2Ucs and Ucs2Asc methods really aren't necessary anymore.
Specifically $.Name is maintained for you, so just dump in UTF-8 to the
attribute, and the module translates in/out of UCS2/UTF-16LE.

I'll document this as well, but this means you could pass in characters
that don't exist in ASCII to the $.Name variable. If this is actually
a problem (I.E. Excel/whatever can't handle these values,) then they can
be filtered later on.